some tweaks to the editor
his makes the changes that need to be made over here for
https://github.com/mitodl/open-discussions/issues/1583 (some initial
tweaks we want to make to the editor's behavior and appearance in
open-discussions).

it also addresses this issue about removing the image caption plugin:
https://github.com/mitodl/open-discussions/issues/1572. We need to
remove this because the plugin doesn't yet properly support read-only
mode.

these changes include:

- an empty editor now has a `.ck-editor__is-empty` class on it
- no more image captions
- no more table insertion options
- changes (specified in
https://github.com/mitodl/open-discussions/issues/1583) to the toolbar
and the block toolbar are made

I think that's about it. the other changes will be made over in the open repo.

pr #7